+++ src/Pages/Proposal/Components/index.js
@@ -1,2 +1,4 @@
 export { default as ProposalDeposits } from './ProposalDeposits';
 export { default as ProposalInformation } from './ProposalInformation';
+export { default as ProposalTimingProgressBar } from './ProposalTimingProgressBar';
+export { default as ProposalTimingTable } from './ProposalTimingTable';